requirements.txt
# Core packages
pandas>=1.5.0
numpy>=1.23.0
plotly
scipy
# Snowflake connector
snowflake-connector-python>=3.0.0
python-dotenv>=1.0.0

# Jupyter for EDA
jupyterlab>=3.0.0
ipykernel>=6.0.0

# (Optional) Visualization libraries if you want to extend EDA
matplotlib>=3.7.0
seaborn>=0.12.0

# (Optional) Logging and progress display
tqdm>=4.65.0
